Extract QuicklinkCoordinator and SnippetExpansionCoordinator
The two largest orchestration blocks in AppCore, ~345 lines together.
Both take their collaborators in init rather than reaching AppCore.shared,
which makes them the first orchestration code a harness could reach.
AppCore keeps thin forwarders so no view changes yet. Bodies moved
verbatim; the open/expand funnels and their gating order are unchanged.
