+++ docs/refactor/ROADMAP.md
@@ -204,7 +204,7 @@ Update this table as phases land. `Blocked` requires a note in the phase's progr
 | 21  | Complete    | `refactor/21-screens-quicklinks-and-emoji` | 2026-08-05 | **The protocol's one navigation hook.** `move(_:axis:from:)` defaults to `nil`, which lets `RootPaletteView` keep its existing per-axis default — linear vertically, caret horizontally — so `EmojiScreen` is the only type that consumes ←/→. ⌘P, ⌘⌫ and ⌥↵ have no member (the gate allows one) and route by `as?` downcast, which phase 23 deletes. `RootPaletteView` 1066 → 982 (−84); 9 files, +434/−231 against an expected 7 and +260/−220 — `docs/development.md` carries the harness's new `EmojiGridGeometry.swift` input. `palette-selection-test` 1,007 → **111,066** assertions, and the **deliberate-break proof passed** only after the column invariant was tightened from `<=` to an exact expected column (4,044 failures, exit 1). **`regression.md` not run** — waived, so AC3–AC7 are structural only. |
 | 22  | Complete    | `refactor/22-screens-clipboard-and-calculator-history` | 2026-08-05 | **The calc card is a row, not an offset.** `rows` is `[.calc(result)] + entries.map(Row.entry)` and every lookup goes through `row(at:)`, so `calcCount` is absent from both new files — the pattern phase 23 reuses. The follow key still reads `store.items.first?.id`; its `vm.mode` guard is now structural (the handler only exists while the screen is mounted) and `scroll` is injected as a `scrollToFollow` closure. One new defaulted protocol member, `hasPrimaryAction(at:)` — without it the error card's "no pill, no ⌘K" rule regresses, since `calcResult` narrows to the launcher here. ⌘P/⌘⌫ route by `as?` downcast like phase 21; `delete` returns `Void` (the screen owns ⌘⌫ even with nothing to delete) while `pin` returns `Bool`. `RootPaletteView` 982 → 832 (−150); 7 files, +424/−292 against an expected 7 and +330/−290. `palette-selection-test` 111,066 → **111,108**, deliberate-break proof passed (exit 1). **`regression.md` waived** — AC4–AC7 and AC10 structural only, including the query-then-copy check. **Unblocks phase 23.** |
 | 23  | Complete    | `refactor/23-screen-launcher-and-offset-collapse` | 2026-08-05 | **M3 closes.** Every mode is a `PaletteScreen`; `calcCount` greps empty and `switch vm.mode` drops 3 → 1 — the survivor is the factory, now non-optional and exhaustive, which deletes the `if let screen` fork at every call site. `LauncherView.swift` is **absent from the diff**, so the nine-section table and its type annotation are byte-identical by construction. `selectionIsRunning` deliberately stays `@State` on the view: a screen has no storage, and reading `runningApps` from `actions(at:)` would put `RunningAppsMonitor` on the palette's body. ⌃⇧Q and ⌘1–⌘5 route by `as?` downcast — safe, not just equivalent, since `paletteIsCollapsed` already implies launcher mode. **AC2 not met** — 665 lines, not ~400: the doc's target assumed a 1126-line file, and the 534 that remain are all things the phase says stay, plus 128 lines of palette chrome **no phase has a home for** (a gap in 29, raised in `progress/23`). Diff +292/−199 against an expected +450/−800 — not net negative, because the file started at 832, not 1126. Debug **and Release** clean (37 s, annotation holds), 0 → 0 warnings, all 17 harnesses green, `palette-selection-test` 111,684. **`regression.md` not run** — waived, so AC4–AC10 are structural only, including the four ↑/↓ walks the risk register calls the real gate. **Unblocks M4.** |
-| 24  | Not started |                                            |            |                                                                                                                                                                                                                                        |
+| 24  | Complete    | `refactor/24-quicklink-and-snippet-coordinators` | 2026-08-05 | **M4 opens.** The two largest orchestration blocks leave `AppCore`: 1342 → 1021 (**−321**, not −345 — 345 lines of body left, ~24 came back as the 13 forwarders and the 2 façade methods the phase itself requires). Both hard gates hold: `AppCore.shared` greps empty in both new files, `DialogController()` stays **1**. **AC3 was proven by direct body diff, not by `git diff -M`** — rename detection can't pair these bodies while `AppCore.swift` still exists, so the pre-change blocks were extracted from `HEAD` and diffed line-by-line; every difference is a receiver rename or `private func` → `func`. Both gating orders byte-identical. **The "+2 properties" budget is short by two**: `reportFailure` and `showMessage` were needed because `dialogs`/`messageHUD` are private and a coordinator may not own its own — phase 25 hits the same wall with `volumeHUD.show` and `dialogs.pickVolume`, neither of which it names. The two coordinators use **different injection shapes on purpose** — `SnippetExpansionCoordinator` holds no `AppCore` reference at all (one `showMessage` closure), `QuicklinkCoordinator` takes `unowned core` for six distinct capabilities. Debug **and** Release clean, 0 → 0 warnings, all 17 harnesses green including both phase gates. **`regression.md` waived** — so AC4–AC6 are structural only, including the "feature off, press the bound shortcut" funnel test the phase calls its single most valuable check. **Unblocks phase 25.** |
 | 25  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 26  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 27  | Not started |                                            |            |                                                                                                                                                                                                                                        |
